Updated create model functionality
